feat(LoRaPlat): v3.4.6 核心架构重构与低功耗机制落地
本次提交对 LoRaPlat 进行了深度的架构优化，重点解决了低功耗唤醒、序列号回绕风险及多节点并发冲突问题，并完善了掉电存储闭环。

[核心变更]
1. 架构重构 (Architecture):
   - Port 层重构为 BSP 角色，新增 `Notify/Check` 机制，解耦中断与业务逻辑。
   - Service 层新增 `CanSleep` 接口，聚合 Manager/Driver/Port 三层状态，实现精准的 Tickless 休眠决策。
   - 移除 Port 层对 MCU 休眠 (`__WFI`) 的直接控制，将调度权归还给 App/Main。

2. 协议升级 (Protocol):
   - 序列号 (Seq) 从 8-bit 升级为 16-bit，解决高速通信下的回绕去重失效问题。

3. 稳定性优化 (Stability):
   - Manager FSM 新增随机退避机制 (CSMA/CA 模拟)，利用物理熵源解决多节点 ACK 死锁。
   - 发送队列深度调整为 2，平衡突发能力与 RAM 占用。
   - 优化 `TxRequest_t` 结构体，移除 Payload 副本，大幅降低静态 RAM 占用 (~800B)。

4. 功能补全 (Features):
   - 修复掉电存储逻辑闭环：在 `_Service_DoReinit` 中增加从 Flash 加载配置的步骤。
   - 增加 Flash 数据 Magic 校验，防止加载无效配置。

5. 代码清理 (Refactor):
   - 调整 `lora_manager_fsm.c` 内部函数顺序，消除隐式声明警告。
   - 统一 STM32 与 ESP32 的 Port 层接口定义。

[影响范围]
- 协议格式变更（不兼容旧版固件）。
- Port 层接口变更（需更新用户侧 ISR 调用）。
- Service 层初始化流程变更。

[后续计划]
- 编写 V1.0 发行版文档。
- Driver 层异步化改造。

+++ LoRaPlatForESP32S3/components/loraplat/src/3_Manager/lora_manager_protocol.h
@@ -47,7 +47,7 @@ typedef struct {
     uint16_t SourceID;       // 源 ID (发送方)
     
     // --- 序号与负载 ---
-    uint8_t  Sequence;       // 包序号
+    uint16_t  Sequence;       // 包序号
     uint8_t  PayloadLen;     // 负载长度
     uint8_t  Payload[LORA_MAX_PAYLOAD_LEN]; // 负载数据
